requirements.txt
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
pillow>=10.2.0
scikit-learn>=1.4.0
torch>=2.1.0
torchvision>=0.16.0
tqdm>=4.66.0